bearer token is being retrieved by auth
getGames are now getting pokemon games by the body but its not dynamic we need to work on that.

the tokens you can put in savely through the env without being scared of uploading them

The apicall should be more like an adapter instead it has been programmed to work for specicly amazon.